requirements.txt
# Core data science
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
openpyxl==3.1.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard
streamlit==1.29.0

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# MLOps / tracking
mlflow==2.9.2

# Data storage
pyarrow==14.0.1
fastparquet==2023.10.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
joblib==1.3.2
